CG-0MQSF0385003SXHV: Fix tutorial E2E resource exhaustion by splitting into per-file vitest runs
Root cause: Phaser 4 RC's CanvasPool accumulates internal canvases
(textures, render targets) across game create/destroy cycles that
can't be freed from JS. After ~8 cycles the browser's per-origin
canvas context limit is exhausted and new games fail silently.

Fix: Split main-street-tutorial-e2e.browser.test.ts into 6 parts
(1-2 tests each) and run each part in a separate vitest invocation.
Each invocation spawns a fresh Chromium instance with an independent
context budget. A wrapper script (scripts/run-tutorial-tests.sh)
orchestrates the sequential runs.

Changes:
- Extract shared helpers: tests/helpers/main-street-tutorial-e2e.ts
- Split tests into 6 part files with 1-6 tests each
- Add scripts/run-tutorial-tests.sh for sequential execution
- Update vite.config.ts: exclude tutorial from main browser project
- Update package.json: run tutorial tests via wrapper after browser
- Fix pre-existing test: T7 requiredCardId assertion (undefined, not
  'evt-grand-opening-15')

Also includes:
- CanvasPool drain in destroyGame for in-process cleanup
- Canvas context validation in bootGameWithTutorial
- Tests 1-12 all pass in their respective vitest runs
