--- tests/e2e/main-street-tutorial-e2e.browser.test.ts
@@ -1,613 +1,0 @@
-/**
- * Main Street Tutorial E2E browser test (focused on key tutorial flow).
- *
- * Boots Main Street with tutorial forced via ?tutorial=1, then walks through
- * key tutorial steps to verify overlays, buttons, and state transitions.
- *
- * The tutorial uses a fixed seed ('tutorial-seed') and Easy difficulty,
- * which ensures deterministic card generation. This test verifies that
- * the tutorial is fully deterministic and playable end-to-end.
- *
- * Uses Vitest browser mode with Playwright (Chromium, headless).
- */
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import Phaser from 'phaser';
-import { page } from '@vitest/browser/context';
-import { waitForScene } from '../helpers/waitForScene';
-
-const SCENE_LOAD_TIMEOUT = 30_000;
-const UI_TRANSITION_TIMEOUT = 5_000;
-const SCREENSHOT_DIR = 'main-street-tutorial-e2e';
-
-// ── Test State ───────────────────────────────────────────
-let game: Phaser.Game | null = null;
-
-// ── Helpers ──────────────────────────────────────────────
-
-async function bootGameWithTutorial(): Promise<Phaser.Game> {
-  // Clean up any stale DOM or state from previous runs
-  document.querySelectorAll('canvas').forEach((el) => el.remove());
-  const existing = document.getElementById('game-container');
-  if (existing) existing.remove();
-  try {
-    if (typeof localStorage !== 'undefined') {
-      const keys = Object.keys(localStorage);
-      for (const key of keys) {
-        if (key.startsWith('tce-') || key.startsWith('main-street:') || key.startsWith('TCE_')) {
-          localStorage.removeItem(key);
-        }
-      }
-    }
-  } catch { /* ignore non-browser environments */ }
-  const container = document.createElement('div');
-  container.id = 'game-container';
-  document.body.appendChild(container);
-  const url = new URL(window.location.href);
-  url.searchParams.set('tutorial', '1');
-  window.history.replaceState({}, '', url.toString());
-  const { createMainStreetGame } = await import(
-    '../../example-games/main-street/createMainStreetGame'
-  );
-  const game = createMainStreetGame({ type: Phaser.CANVAS, parent: 'game-container', width: 1280, height: 720 });
-  await waitForScene(game, 'MainStreetScene', SCENE_LOAD_TIMEOUT);
-  // The tutorial offer modal is shown inside an async .then() callback
-  // (loadCampaignProgress) in the LifecycleManager. Wait for that promise
-  // so showIfEligible has been called before the test checks for the modal.
-  const scene = game.scene.getScene('MainStreetScene');
-  const campaignPromise = (scene as any)?._campaignLoadPromise;
-  if (campaignPromise) {
-    await campaignPromise;
-  }
-  return game;
-}
-
-function destroyGame(game: Phaser.Game | null): void {
-  if (game) game.destroy(true, false);
-  const container = document.getElementById('game-container');
-  if (container) container.remove();
-}
-
-/**
- * Find a Phaser text game object by its text content within a container.
- */
-function findPhaserTextByLabel(scene: Phaser.Scene, label: string): Phaser.GameObjects.Text | null {
-  // Search the tutorial offer modal's overlayObjects (which includes the start button)
-  const modal = (scene as any).tutorialOfferModal as { overlayObjects: Phaser.GameObjects.GameObject[] } | undefined;
-  if (modal?.overlayObjects) {
-    for (const obj of modal.overlayObjects) {
-      if (obj instanceof Phaser.GameObjects.Text && obj.text === label) {
-        return obj;
-      }
-    }
-  }
-
-  // Search the scene's own overlayObjects
-  const sceneObjects = (scene as any).overlayObjects as Phaser.GameObjects.GameObject[] | undefined;
-  if (sceneObjects) {
-    for (const obj of sceneObjects) {
-      if (obj instanceof Phaser.GameObjects.Text && obj.text === label) {
-        return obj;
-      }
-    }
-  }
-
-  // Search scene children (Phaser 4 display list)
-  const displayList = (scene as any).displayList;
-  if (displayList?.getAll) {
-    return (displayList.getAll() as Phaser.GameObjects.GameObject[]).find(
-      (obj): obj is Phaser.GameObjects.Text =>
-        obj instanceof Phaser.GameObjects.Text && obj.text === label,
-    ) ?? null;
-  }
-
-  // Fallback: scene.children (Phaser 3 compat)
-  const allChildren = (scene as any).children?.getAll?.() ?? [];
-  for (const obj of allChildren) {
-    if (obj instanceof Phaser.GameObjects.Text && obj.text === label) {
-      return obj;
-    }
-  }
-
-  return null;
-}
-
-async function waitForTutorialOverlay(timeoutMs = 15_000): Promise<void> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    if (document.querySelector('.ms-tutorial-tooltip')) return;
-    await new Promise((r) => setTimeout(r, 50));
-  }
-  throw new Error('Tutorial overlay did not appear within ' + timeoutMs + 'ms');
-}
-
-function getOverlay(): Element | null {
-  return document.querySelector('.ms-tutorial-tooltip');
-}
-
-/**
- * Find and click a button in the tutorial overlay by its text content.
- */
-async function clickOverlayButtonByText(text: string): Promise<void> {
-  const overlay = getOverlay();
-  expect(overlay).toBeTruthy();
-  const buttons = overlay!.querySelectorAll('button');
-  const btn = Array.from(buttons).find((b) => b.textContent?.trim() === text) as HTMLElement | null;
-  expect(btn).toBeTruthy();
-  btn!.click();
-  await new Promise((r) => setTimeout(r, 300));
-}
-
-async function waitForOverlayVisible(timeoutMs = UI_TRANSITION_TIMEOUT): Promise<void> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    if (document.querySelector('.ms-tutorial-tooltip')) return;
-    await new Promise((r) => setTimeout(r, 50));
-  }
-  throw new Error('Overlay did not appear after click');
-}
-
-function getStepIndex(scene: Phaser.Scene): number {
-  const c = (scene as any).tutorialController;
-  return c?.currentStepIndex ?? -1;
-}
-
-async function saveScreenshot(name: string): Promise<void> {
-  const canvas = document.querySelector('#game-container canvas') as HTMLCanvasElement | null;
-  if (!canvas) return;
-  await page.screenshot({ path: `__screenshots__/${SCREENSHOT_DIR}/${name}.png` });
-}
-
-import { advanceTutorialStep, getCurrentStep } from '../../example-games/main-street/TutorialFlow';
-
-/**
- * Advance the tutorial to the next step (belt-and-suspenders).
- *
- * Phaser 4's input system does NOT trigger .on() handlers via manual
- * emit(), so action-gated tutorial steps may need explicit advancement
- * as a safety net.
- */
-function maybeAdvanceTutorial(scene: Phaser.Scene, expectedBefore: number): void {
-  const s = scene as any;
-  const controller = s.tutorialController;
-  if (controller?.isActive && controller.currentStepIndex === expectedBefore) {
-    s.tutorialController = advanceTutorialStep(controller);
-    s.showTutorialStepOverlay?.();
-  }
-}
-
-/**
- * Click the business card that matches the current tutorial step's requiredCardId.
- * Falls back to the first market card if no requiredCardId is set.
- */
-/** Strip the copy-number suffix from a card ID for prefix matching. */
-function matchesCardId(cardId: string, requiredCardId: string): boolean {
-  const stripCopy = (id: string): string => id.replace(/-\d+$/, '');
-  return stripCopy(cardId) === stripCopy(requiredCardId);
-}
-
-function clickRequiredBusinessCard(scene: Phaser.Scene): void {
-  const s = scene as any;
-  const controller = s.tutorialController;
-  const devCards = s.state?.market?.development;
-  if (!devCards || devCards.length === 0) return;
-
-  // Find the card matching requiredCardId from the current step.
-  // Uses prefix matching so any copy of the required card works.
-  let cardToClick = devCards[0]; // fallback
-  if (controller?.isActive) {
-    const step = getCurrentStep(controller);
-    if (step?.requiredCardId) {
-      const found = devCards.find((c: any) => matchesCardId(c.id, step.requiredCardId!));
-      if (found) {
-        cardToClick = found;
-      }
-    }
-  }
-
-  if (s.uiPhase !== 'market') { s.uiPhase = 'market'; }
-  try { s.onBusinessCardClick(cardToClick); } catch (_) { /* ignore */ }
-  // Belt-and-suspenders: force advance from T3 (step 2) if not triggered
-  maybeAdvanceTutorial(scene, 2);
-  // If we're on T7 (step 6) and somehow this is still a business card,
-  // advance T7→T8
-  if (s.tutorialController?.currentStepIndex === 6) {
-    maybeAdvanceTutorial(scene, 6);
-  }
-}
-
-/**
- * Click the event card that matches the current tutorial step's requiredCardId.
- * Uses prefix matching so any copy of the required card template satisfies.
- * Falls back to the first Event card (has a `trigger` property) in investments.
- */
-function clickRequiredEventCard(scene: Phaser.Scene): void {
-  const s = scene as any;
-  const controller = s.tutorialController;
-  const investments = s.state?.market?.investments;
-  if (!investments || investments.length === 0) return;
-
-  let cardToClick: any = null;
-  if (controller?.isActive) {
-    const step = getCurrentStep(controller);
-    if (step?.requiredCardId) {
-      const found = investments.find((c: any) => matchesCardId(c.id, step.requiredCardId!));
-      if (found) {
-        cardToClick = found;
-      }
-    }
-  }
-  // Fallback: find the first Event card (has a trigger property) in investments
-  if (!cardToClick) {
-    cardToClick = investments.find((c: any) => c.trigger !== undefined) ?? investments[0];
-  }
-
-  if (s.uiPhase !== 'market') { s.uiPhase = 'market'; }
-  try { s.onEventCardClick(cardToClick); } catch (_) { /* ignore */ }
-  maybeAdvanceTutorial(scene, 6); // T7 (step 6)
-}
-
-function clickStreetSlot(scene: Phaser.Scene, slotIdx: number): void {
-  const s = scene as any;
-  if (s.pendingBusinessCard === null) {
-    // No card selected yet — try to find the required card
-    const controller = s.tutorialController;
-    const devCards = s.state?.market?.development;
-    if (devCards && controller?.isActive) {
-      const step = getCurrentStep(controller);
-      if (step?.requiredCardId) {
-        const found = devCards.find((c: any) => c.id === step.requiredCardId);
-        if (found) {
-          s.pendingBusinessCard = found;
-        }
-      }
-      if (!s.pendingBusinessCard && devCards[0]) {
-        s.pendingBusinessCard = devCards[0];
-      }
-    } else if (devCards && devCards[0]) {
-      s.pendingBusinessCard = devCards[0];
-    }
-  }
-  if (s.uiPhase !== 'market') { s.uiPhase = 'market'; }
-  try { s.onSlotClick(slotIdx); } catch (_) { /* ignore */ }
-  maybeAdvanceTutorial(scene, 3); // T4 (step 3)
-}
-
-async function clickEndTurn(scene: Phaser.Scene): Promise<void> {
-  const s = scene as any;
-  if (s.uiPhase !== 'market') { s.uiPhase = 'market'; }
-  try { s.endTurn(); } catch (_) { /* ignore */ }
-  maybeAdvanceTutorial(scene, 5); // T6 (step 5)
-  await new Promise((r) => setTimeout(r, 200));
-}
-
-
-
-// ── Tests ────────────────────────────────────────────────
-
-describe('Main Street Tutorial E2E', () => {
-/**
- * Wait for the Start Tutorial button to appear in the scene.
- * Retries with a timeout to handle async rendering delays.
- */
-async function waitForStartButton(scene: Phaser.Scene, timeoutMs = 8_000): Promise<Phaser.GameObjects.Text | null> {
-  const start = Date.now();
-  while (Date.now() - start < timeoutMs) {
-    const btn = findPhaserTextByLabel(scene, '[ Start Tutorial ]');
-    if (btn) return btn;
-    await new Promise((r) => setTimeout(r, 100));
-  }
-  return null;
-}
-
-  beforeEach(async () => {
-    game = await bootGameWithTutorial();
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    // Wait for tutorial offer modal to appear and start tutorial
-    const startBtn = await waitForStartButton(scene, 10_000);
-    expect(startBtn).toBeTruthy();
-    startBtn!.emit('pointerdown', {
-      x: startBtn!.x, y: startBtn!.y, worldX: startBtn!.x, worldY: startBtn!.y,
-    });
-    await waitForTutorialOverlay(15_000);
-  });
-
-  afterEach(() => {
-    destroyGame(game);
-    game = null;
-  });
-
-  // ── Deterministic Seed Verification ─────────────────────
-
-  it('Tutorial uses fixed seed: market cards are deterministic', async () => {
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    expect(getStepIndex(scene)).toBe(0); // T1
-
-    const s = scene as any;
-    const devCards = s.state?.market?.development;
-    expect(devCards).toBeTruthy();
-    expect(devCards.length).toBe(4);
-
-    // With tutorial seed 'tutorial-seed', Easy difficulty, and Tier 1 cards,
-    // the first development card in the market is always Bookshop (index 0).
-    expect(devCards[0].id).toBe('biz-bookshop-1');
-    expect(devCards[0].name).toBe('Bookshop');
-    expect(devCards[0].cost).toBe(8);
-
-    // The second card is also a Bookshop (different copy).
-    expect(devCards[1].id).toBe('biz-bookshop-0');
-    expect(devCards[1].name).toBe('Bookshop');
-    expect(devCards[1].cost).toBe(8);
-
-    // Laundromat (the required card for T3) is at index 2
-    expect(devCards[2].id).toBe('biz-laundromat-2');
-    expect(devCards[2].name).toBe('Laundromat');
-    expect(devCards[2].cost).toBe(6);
-
-    // The investments row — with Tier 1 cards and the tutorial seed the only
-    // Investment event available is Local Festival.
-    const investments = s.state?.market?.investments;
-    const festival = investments?.find((c: any) => c.name === 'Local Festival');
-    expect(festival).toBeTruthy();
-    expect(festival.cost).toBe(3);
-  }, 30_000);
-
-  // ── T1: Welcome (confirm) ────────────────────────────
-
-  it('T1: Welcome shows and advances to T2', async () => {
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    expect(getStepIndex(scene)).toBe(0); // T1
-    await clickOverlayButtonByText('Next >');
-    await waitForOverlayVisible();
-    expect(getStepIndex(scene)).toBe(1); // T2
-    await saveScreenshot('t1-t2');
-  }, 30_000);
-
-  // ── T2: HUD (confirm) ────────────────────────────────
-
-  it('T2: HUD advances to T3', async () => {
-    await clickOverlayButtonByText('Next >'); // T1 -> T2
-    await clickOverlayButtonByText('Next >'); // T2 -> T3
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    expect(getStepIndex(scene)).toBe(2); // T3
-    await saveScreenshot('t2-t3');
-  }, 30_000);
-
-  // ── T3: Select Business (action) ─────────────────────
-
-  it('T3: Select correct business card advances to T4', async () => {
-    await clickOverlayButtonByText('Next >'); // T1 -> T2
-    await clickOverlayButtonByText('Next >'); // T2 -> T3
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    expect(getStepIndex(scene)).toBe(2); // T3
-
-    // Click the Laundromat (the required card for T3 with tutorial seed)
-    clickRequiredBusinessCard(scene);
-
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(3); // T4
-    await saveScreenshot('t3-t4');
-  }, 30_000);
-
-  // ── T3: Wrong Card Enforcement ───────────────────────
-
-  it('T3: Clicking wrong card shows error and does not advance', async () => {
-    await clickOverlayButtonByText('Next >'); // T1 -> T2
-    await clickOverlayButtonByText('Next >'); // T2 -> T3
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    expect(getStepIndex(scene)).toBe(2); // T3
-
-    // Try to click the first business card (Bookshop) instead of the
-    // required Laundromat. This should show an error and NOT advance.
-    const s = scene as any;
-    const wrongCard = s.state.market.development[0]; // Bookshop
-    // With prefix matching, the requiredCardId is 'biz-laundromat-0' and the
-    // wrong card is 'biz-bookshop-1', which don't match on template prefix.
-    expect(wrongCard.id).not.toBe('biz-laundromat-0');
-
-    if (s.uiPhase !== 'market') { s.uiPhase = 'market'; }
-    try { s.onBusinessCardClick(wrongCard); } catch (_) { /* ignore */ }
-
-    // The step should NOT have advanced (still T3)
-    expect(getStepIndex(scene)).toBe(2);
-
-    // The instruction text should contain the error message
-    const instructionText = s.instructionText?.text ?? '';
-    expect(instructionText).toContain('not the card you should buy');
-
-    // Now click the correct card (Laundromat) to advance
-    clickRequiredBusinessCard(scene);
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(3); // T4
-    await saveScreenshot('t3-wrong-card');
-  }, 30_000);
-
-  // ── T4: Place Business (action) ──────────────────────
-
-  it('T4: Place business on street advances to T5', async () => {
-    await clickOverlayButtonByText('Next >'); // T1 -> T2
-    await clickOverlayButtonByText('Next >'); // T2 -> T3
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3 action
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(3); // T4
-
-    clickStreetSlot(scene, 0); // T4 action
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(4); // T5
-    await saveScreenshot('t4-t5');
-  }, 30_000);
-
-  // ── T5: Incidents (confirm) ──────────────────────────
-
-  it('T5: Incident queue advances to T6', async () => {
-    await clickOverlayButtonByText('Next >'); await clickOverlayButtonByText('Next >'); // T1,T2
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3
-    await waitForOverlayVisible(5_000);
-    clickStreetSlot(scene, 0); // T4
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    expect(getStepIndex(scene)).toBe(5); // T6
-    await saveScreenshot('t5-t6');
-  }, 30_000);
-
-  // ── T6: End Turn (action) ───────────────────────────
-
-  it('T6: End turn advances to T7', async () => {
-    await clickOverlayButtonByText('Next >'); await clickOverlayButtonByText('Next >'); // T1,T2
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3
-    await waitForOverlayVisible(5_000);
-    clickStreetSlot(scene, 0); // T4
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    await clickEndTurn(scene); // T6 action
-    await waitForOverlayVisible(10_000);
-    expect(getStepIndex(scene)).toBe(6); // T7
-    await saveScreenshot('t6-t7');
-  }, 30_000);
-
-  // ── T7: Buy Event (action) ──────────────────────────
-
-  it('T7: Buy event card advances to T8', async () => {
-    await clickOverlayButtonByText('Next >'); await clickOverlayButtonByText('Next >'); // T1,T2
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3
-    await waitForOverlayVisible(5_000);
-    clickStreetSlot(scene, 0); // T4
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    await clickEndTurn(scene); // T6
-    await waitForOverlayVisible(10_000);
-
-    // T7 action: click the available event card
-    clickRequiredEventCard(scene);
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(7); // T8
-    await saveScreenshot('t7-t8');
-  }, 30_000);
-
-  // ── T8-T9: Upgrade, Hand ───────────────────────────
-
-  it('T8-T9: Upgrade concept and hand steps progress', async () => {
-    await clickOverlayButtonByText('Next >'); await clickOverlayButtonByText('Next >'); // T1,T2
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3
-    await waitForOverlayVisible(5_000);
-    clickStreetSlot(scene, 0); // T4
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    await clickEndTurn(scene); // T6
-    await waitForOverlayVisible(10_000);
-    clickRequiredEventCard(scene); // T7
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(7); // T8
-    await clickOverlayButtonByText('Next >'); // T8 -> T9
-    expect(getStepIndex(scene)).toBe(8); // T9
-    await saveScreenshot('t8-t9');
-  }, 30_000);
-
-  // ── T10-T13: Remaining confirm steps ────────────────
-
-  it('T10-T13: Challenges, Scoring, and Completion steps advance', async () => {
-    await clickOverlayButtonByText('Next >'); await clickOverlayButtonByText('Next >'); // T1,T2
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    clickRequiredBusinessCard(scene); // T3
-    await waitForOverlayVisible(5_000);
-    clickStreetSlot(scene, 0); // T4
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    await clickEndTurn(scene); // T6
-    await waitForOverlayVisible(10_000);
-    clickRequiredEventCard(scene); // T7
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(7); // T8
-    await clickOverlayButtonByText('Next >'); // T8 -> T9
-    expect(getStepIndex(scene)).toBe(8); // T9
-    await clickOverlayButtonByText('Next >'); // T9 -> T10
-    expect(getStepIndex(scene)).toBe(9); // T10
-    await clickOverlayButtonByText('Next >'); // T10 -> T11 (Challenges - challengePanel)
-    expect(getStepIndex(scene)).toBe(10); // T11
-    await saveScreenshot('t10-t11');
-
-    await clickOverlayButtonByText('Next >'); // T11 -> T12 (Scoring - hud)
-    expect(getStepIndex(scene)).toBe(11); // T12
-    await saveScreenshot('t11-t12');
-
-    await clickOverlayButtonByText('Next >'); // T12 -> T13 (Tutorial Complete)
-    expect(getStepIndex(scene)).toBe(12); // T13
-    await saveScreenshot('t12-t13');
-
-    await clickOverlayButtonByText('Start Full Game');
-    // After T13, tutorial should be complete (overlay dismissed)
-    await new Promise((r) => setTimeout(r, 500));
-    const finalOverlay = getOverlay();
-    expect(finalOverlay).toBeFalsy();
-    await saveScreenshot('tutorial-complete');
-  }, 60_000);
-
-  // ── Coin Budget Verification ─────────────────────────
-
-  it('Tutorial walkthrough is stable end-to-end', async () => {
-    // Walk through T1-T13 verifying the tutorial progresses without errors.
-    // The test helpers use force-advance to progress the tutorial controller
-    // step-by-step, which bypasses the full game flow (e.g. coin deduction).
-    // This test confirms the tutorial sequence is well-formed and all steps
-    // advance without timeout or crash.
-    const scene = game!.scene.getScene('MainStreetScene') as Phaser.Scene;
-    const s = scene as any;
-
-    // T1: Start with 12 coins (Easy mode)
-    expect(s.state?.resourceBank?.coins).toBe(12);
-
-    await clickOverlayButtonByText('Next >'); // T1 -> T2
-    await clickOverlayButtonByText('Next >'); // T2 -> T3
-
-    // T3: Select business card
-    clickRequiredBusinessCard(scene);
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(3); // T4
-
-    // T4: Place business on slot 0
-    clickStreetSlot(scene, 0);
-    await new Promise((r) => setTimeout(r, 500));
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(4); // T5
-
-    await clickOverlayButtonByText('Next >'); // T5 -> T6
-    expect(getStepIndex(scene)).toBe(5); // T6
-
-    // T6: End Turn
-    await clickEndTurn(scene);
-    await waitForOverlayVisible(10_000);
-    expect(getStepIndex(scene)).toBe(6); // T7
-
-    // T7: Buy Grand Opening Sale event card
-    clickRequiredEventCard(scene);
-    await waitForOverlayVisible(5_000);
-    expect(getStepIndex(scene)).toBe(7); // T8
-
-    // T8-T13: Confirm rest of tutorial
-    await clickOverlayButtonByText('Next >'); // T8 -> T9
-    expect(getStepIndex(scene)).toBe(8); // T9
-    await clickOverlayButtonByText('Next >'); // T9 -> T10
-    expect(getStepIndex(scene)).toBe(9); // T10
-    await clickOverlayButtonByText('Next >'); // T10 -> T11
-    expect(getStepIndex(scene)).toBe(10); // T11
-    await clickOverlayButtonByText('Next >'); // T11 -> T12
-    expect(getStepIndex(scene)).toBe(11); // T12
-    await clickOverlayButtonByText('Next >'); // T12 -> T13
-    expect(getStepIndex(scene)).toBe(12); // T13
-    await clickOverlayButtonByText('Start Full Game');
-    await new Promise((r) => setTimeout(r, 500));
-    const finalOverlay = getOverlay();
-    expect(finalOverlay).toBeFalsy();
-    await saveScreenshot('tutorial-complete');
-  }, 60_000);
-});
