+++ tests/main-street/tutorial-flow.test.ts
@@ -44,7 +44,7 @@ describe('UNIFIED_TUTORIAL_STEPS', () => {
   it('T3 is action gate with select-business requiredAction and requiredCardId', () => { const t=findStep('T3'); expect(t.gate).toBe('action'); expect(t.requiredAction).toBe('select-business'); expect(t.requiredCardId).toBe('biz-laundromat-0'); expect(t.highlightZone).toBe('marketBusinessRow'); });
   it('T4 is action gate with place-business requiredAction and no requiredCardId', () => { const t=findStep('T4'); expect(t.gate).toBe('action'); expect(t.requiredAction).toBe('place-business'); expect(t.requiredCardId).toBeUndefined(); expect(t.highlightZone).toBe('streetGrid'); });
   it('T6 is action gate with end-turn requiredAction and no requiredCardId', () => { const t=findStep('T6'); expect(t.gate).toBe('action'); expect(t.requiredAction).toBe('end-turn'); expect(t.requiredCardId).toBeUndefined(); expect(t.highlightZone).toBe('endTurnButton'); });
-  it('T7 is action gate with buy-event requiredAction and requiredCardId', () => { const t=findStep('T7'); expect(t.gate).toBe('action'); expect(t.requiredAction).toBe('buy-event'); expect(t.requiredCardId).toBe('evt-grand-opening-15'); expect(t.highlightZone).toBe('investmentsRow'); });
+  it('T7 is action gate with buy-event requiredAction and no requiredCardId', () => { const t=findStep('T7'); expect(t.gate).toBe('action'); expect(t.requiredAction).toBe('buy-event'); expect(t.requiredCardId).toBeUndefined(); expect(t.highlightZone).toBe('investmentsRow'); });
   it('T8 is confirm gate (upgrade concept reference, not action-gated)', () => { const t=findStep('T8'); expect(t.gate).toBe('confirm'); expect(t.requiredAction).toBeUndefined(); expect(t.highlightZone).toBe('investmentsRow'); });
   it('T13 is confirm gate with completionModal highlight', () => { expect(findStep('T13').gate).toBe('confirm'); expect(findStep('T13').highlightZone).toBe('completionModal'); });
 });
